+++ tests/test_actions.py
@@ -2,6 +2,7 @@ import unittest
 
 import imap_tools
 from tests.utils import MailboxTestCase, TEST_MAILBOX_NAME_SET, get_test_mailbox
+from imap_tools.errors import MailboxCopyError
 
 
 class ActionTest(MailboxTestCase):
@@ -29,7 +30,10 @@ class ActionTest(MailboxTestCase):
 
             # COPY
             mailbox.folder.set(mailbox.folder_test_base)
-            mailbox.copy([msg.uid for msg in mailbox.fetch()], mailbox.folder_test_temp1)
+            uid_set = [msg.uid for msg in mailbox.fetch()]
+            mailbox.copy(uid_set, mailbox.folder_test_temp1)
+            with self.assertRaises(MailboxCopyError):
+                mailbox.copy(uid_set, '__nonexistent_folder__')
             self.assertEqual(len(list(mailbox.fetch())), self.base_test_msg_cnt)
             mailbox.folder.set(mailbox.folder_test_temp1)
             self.assertEqual(len(list(mailbox.fetch())), self.base_test_msg_cnt)
@@ -43,13 +47,13 @@ class ActionTest(MailboxTestCase):
 
             # FLAG
             mailbox.folder.set(mailbox.folder_test_temp2)
-            mailbox.flag([msg.uid for msg in mailbox.fetch()], imap_tools.MessageFlags.FLAGGED, True)
-            self.assertTrue(all([imap_tools.MessageFlags.FLAGGED in msg.flags for msg in mailbox.fetch()]))
+            mailbox.flag([msg.uid for msg in mailbox.fetch()], imap_tools.MailMessageFlags.FLAGGED, True)
+            self.assertTrue(all([imap_tools.MailMessageFlags.FLAGGED in msg.flags for msg in mailbox.fetch()]))
 
             # SEEN
             mailbox.folder.set(mailbox.folder_test_temp2)
             mailbox.seen([msg.uid for msg in mailbox.fetch()], False)
-            self.assertTrue(all([imap_tools.MessageFlags.SEEN not in msg.flags
+            self.assertTrue(all([imap_tools.MailMessageFlags.SEEN not in msg.flags
                                  for msg in mailbox.fetch(mark_seen=False)]))
 
             # DELETE